requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssmin==0.2.0
ffmpeg-python==0.2.0
filelock==3.10.7
Flask==2.2.3
Flask-Assets==2.0
Flask-Cors==3.0.10
Flask-WTF==1.1.1
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.28.2
requests-file==1.5.1
six==1.16.0
SQLAlchemy==2.0.8
tldextract==3.4.0
typing_extensions==4.5.0
urllib3==1.26.15
webassets==2.0
Werkzeug==2.2.3
WTForms==3.0.1
zope.event==4.6
zope.interface==6.0
